api/pyproject.toml
[project]
name = "sangokushi_api"
version = "3.10.11"
description = "Add your description here"
authors = [
    { name = "coji", email = "coji@techtalk.jp" }
]
dependencies = [
    "sentence-transformers>=2.2.2",
    "fugashi[unidic-lite]>=1.2.1",
    "fastapi>=0.101.1",
    "pydantic>=2.2.1",
    "uvicorn[standard]>=0.23.2",
    "torch==2.0.1",
    "hyperdb-python>=0.1.4",
    "scikit-learn>=1.3.0",
    "umap-learn>=0.5.3",
    "llvmlite>=0.40.1",
    "numba>=0.57.1",
    "setuptools>=68.1.2",
    "prisma>=0.10.0",
    "isort>=5.12.0",
    "pandas>=2.0.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=5.0.4",
    "pyproject-flake8>=5.0.4.post1",
    "mypy>=1.5.1",
]

[[tool.rye.sources]]
name = "torch"
url = "https://download.pytorch.org/whl/cpu"
type = "index"

[tool.rye.scripts]
start = { cmd = 'uvicorn server:app --reload --host 0.0.0.0 --port 8000', env = { ENV = 'development' } }
"build:db" = { cmd = 'python -m src.sangokushi_api.commands.build_db'}
finetune = { cmd = 'python -m src.sangokushi_api.commands.finetune'}

[tool.hatch.metadata]
allow-direct-references = true

# 以下各フォーマッター・リンターの詳細設定
[tool.black]
# 最大文字数 指定文字数を超えたら改行
line_length = 79

[tool.flake8]
# blackのline_lengthに合わせてください
max-line-length = 80
# 警告非表示一覧
extend-ignore = ["E203", "W503"]

[tool.isort]
profile = "black"

[tool.mypy]
# エラー時のメッセージを詳細表示
show_error_context = true
# エラー発生箇所の行数/列数を表示
show_column_numbers = true
# サードパーティ製モジュールの型チェックを無視
ignore_missing_imports = true
# 関数の引数/戻り値に型指定必須
disallow_untyped_defs = false
# デフォルト引数にNoneを取る場合、型アノテーションにOptional必須
no_implicit_optional = true
# 戻り値がAny型ではない関数の戻り値がAny型のとき警告
warn_return_any = true
# 冗長なキャストに警告
warn_redundant_casts = true
